added functionality to close drop down on outside click
