+++ client/public/src/components/qtyDropDown.jsx
@@ -1,35 +1,26 @@
 import React, { useEffect, useRef } from 'react';
 import ReactDOM from 'react-dom';
 
-import {QtyDropdownContent} from '../styling/qtyDropdown.jsx';
+import {QtyDropdownContent, QtyBtn, BtnText, DropdownArrow} from '../styling/qtyDropdown.jsx';
 import QtyNumber from './qtyNumber.jsx';
 
-const alertWhenClickedOutside = (ref) => {
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (ref.current && !ref.current.contains(event.target)) {
-        console.log('outside click!');
-      }
-    }
+import alertWhenClickedOutside from '../utils/outsideClick.js'
 
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    }
-  }, [ref]);
-}
 
-const QtyDropDown = ({active}) => {
-
+const QtyDropDown = ({active, toggleClass, closeDropdown}) => {
   // returns range form 0 to max
   const createRange = (max) => {
     return Array.from(Array(99).keys());
   }
 
   const wrapperRef = useRef(null);
-  alertWhenClickedOutside(wrapperRef);
+  alertWhenClickedOutside(wrapperRef, closeDropdown);
   return (
     <div ref={wrapperRef}>
+      <QtyBtn onClick={()=> toggleClass()}>
+        <BtnText>1</BtnText>
+        <DropdownArrow>&#8964;</DropdownArrow>
+      </QtyBtn>
       <QtyDropdownContent active={active}>
           {createRange(99).map( number => <QtyNumber number={number + 1} />)}
         </QtyDropdownContent>
