+++ client/public/src/components/selectQty.jsx
@@ -3,8 +3,7 @@ import ReactDOM from 'react-dom';
 
 import QtyDropDown from './qtyDropdown.jsx';
 
-import {SelectQtyContainer, SelectQtyTitle, QtyBtn, BtnText,
-  DropdownArrow} from '../styling/qtyDropdown.jsx';
+import {SelectQtyContainer, SelectQtyTitle} from '../styling/qtyDropdown.jsx';
 
 
 const SelectQty = () => {
@@ -15,7 +14,9 @@ const SelectQty = () => {
     setDropDownState({active: newState});
   }
 
-
+  const closeDropdown = () => {
+    setDropDownState({active: false});
+  }
 
   useEffect( () => {}, [dropdownState]);
 
@@ -24,11 +25,11 @@ const SelectQty = () => {
   return (
     <SelectQtyContainer>
       <SelectQtyTitle>Quantity</SelectQtyTitle>
-      <QtyBtn onClick={()=> toggleClass()}>
+      {/* <QtyBtn onClick={()=> toggleClass()}>
         <BtnText>1</BtnText>
         <DropdownArrow>&#8964;</DropdownArrow>
-      </QtyBtn>
-        <QtyDropDown active={dropdownState.active}/>
+      </QtyBtn> */}
+        <QtyDropDown active={dropdownState.active} toggleClass={toggleClass} closeDropdown= {closeDropdown}/>
     </SelectQtyContainer>
   )
 }
